Move keyring::store to keyring
